[Feat]: 퀴즈 문제 풀기 구현

+++ app/src/main/java/com/beside/hackathon/data/model/quiz/QuizSubmitRequest.kt
@@ -1,5 +1,5 @@
 package com.beside.hackathon.data.model.quiz
 
 data class QuizSubmitRequest(
-    val selectedOptionIds: List<Long>,
+    val selectedOptionIds: List<Int>,
 )
